% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{campus_crime_hawkes}
\alias{campus_crime_hawkes}
\title{Data regarding campus crime.}
\format{
A data.frame with 20 rows and 4 columns.
\describe{
  \item{crimes}{The number of crimes as applicable to the school}
  \item{police}{Number of police officers at the school}
  \item{enrolled}{Total number of students enrolled at the school}
  \item{private}{Private school (1), and others (0).}
}
}
\source{
\url{https://www.hawkeslearning.com/Statistics/dataset-descriptions.html#campus-crime}
}
\usage{
campus_crime_hawkes
}
\description{
Data regarding campus crime.
}
\keyword{datasets}
