% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{cereal}
\alias{cereal}
\title{Nutrition Information for Cereals}
\format{
A data.frame with 77 rows and 15 variables.
\describe{
  \item{Shelf}{One of Top, Bottom, or Middle, describing the position of the
  cereal on the shelf at the grocery store.}
  \item{Name}{The name of the cereal.}
  \item{Manufacturer}{The manufacturer of the cereal, including American home
  food products (A), General Mills (G), Kelloggs (K), Nabisco (N), Post
  Brands (P), Quaker Oats (Q), and Ralston (R).}
  \item{Type}{One of cold (C) or hot (H).}
  \item{Calories}{The calories in one serving.}
  \item{Protein}{The number of grams of protein in one serving.}
  \item{Fat}{The number of grams of fat in one serving.}
  \item{Sodium}{The number of milligrams (mg) of sodium in one serving.}
  \item{Fiber}{The number of grams of fiber in one serving.}
  \item{Carbohydrates}{The number of grams of carbohydrates in one serving.}
  \item{Sugars}{The numebr of grams of sugar in one serving.}
  \item{Vitamins}{vitamins and minerals - 0, 25, or 100, indicating the
  typical percentage of FDA}
  \item{Weight}{The weight of one serving size of cereal.}
  \item{Cup}{number of cups in one serving}
}
}
\source{
\url{https://www.hawkeslearning.com/Statistics/beg3/datasets.html}

\url{https://www.kaggle.com/jeandsantos/breakfast-cereals-data-analysis-and-clustering}
}
\usage{
cereal
}
\description{
A dataset that provides nutritional information regarding cereals.
}
\keyword{datasets}
