% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_descriptions.R
\docType{data}
\name{colleges}
\alias{colleges}
\title{University Statistics}
\format{
A data frame with 831 rows and 10 columns.
\describe{
  \item{name}{Name of the college or university}
  \item{endowment_begin}{Size of the school endowment at the beginning of
     the fiscal year, in billions of dollars}
  \item{endowment_end}{Size of the school endowment at the end of
     the fiscal year, in billions of dollars}
  \item{enrollment}{Number of enrolled undergraduates}
  \item{enrollment_men_prop}{Proportion of male undergraduates}
  \item{enrollment_women_prop}{Proportion of female undergraduates}
  \item{public}{Is the school a public institution?}
  \item{attendance_cost}{Average cost of attendance, in thousands of dollars}
  \item{income_median}{Median family income, in thousands of dollars}
  \item{earnings}{Median earnings 10 years after enrollment,
    in thousands of dollars}
}
}
\source{
<https://collegescorecard.ed.gov/data/>
}
\usage{
colleges
}
\description{
A subset of data from the U.S. Department of Education's College Score
  card. Contains information about institutions across the US. Institutions
  are only retained if they offer 4-year programs, have more than 2k
  students, have an endowment larger than 10 million dollars, and have
  non-missing information for all the variables included in the dataset. The
  data represent the most recent statistics that were compiled in May
  of 2025.
}
\keyword{datasets}
