% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_descriptions.R
\docType{data}
\name{covid_case}
\alias{covid_case}
\title{Covid Case Data - Country Level Aggregates}
\format{
A data frame with 241 rows and 6 columns.

\describe{
  \item{name}{Country name}
  \item{region}{World Health Organization region}
  \item{case_total}{Cumulative number of cases}
  \item{case_pop}{Cumulative number of cases per 100,000 people}
  \item{death_total}{Cumulative number of deaths}
  \item{death_pop}{Cumulative number of deaths per 100,000 people}
}
}
\source{
<https://data.who.int/dashboards/covid19/>
}
\usage{
covid_case
}
\description{
A snapshot in time of COVID case level aggregates by country, as prepared
  by the World Health Organization.
}
\keyword{datasets}
