% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_descriptions.R
\docType{data}
\name{covid_hos}
\alias{covid_hos}
\title{Covid Case Data - Country Level Time Series Data}
\format{
A data frame with 241 rows and 6 columns.

\describe{
  \item{date}{Date of report, in yyyy-mm-dd format}
  \item{country}{Country name}
  \item{region}{World Health Organization region}
  \item{new_hos}{Number of new hospitalizations reported that week.}
}
}
\source{
<https://data.who.int/dashboards/covid19/>
}
\usage{
covid_hos
}
\description{
Weekly covid case statistics by country, as reported by the World Health
  Organization.
}
\keyword{datasets}
