% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5055_data_descriptions.R
\docType{data}
\name{covid_raw}
\alias{covid_raw}
\title{COVID-19 Case Count Data}
\format{
A data.frame with 333,459 rows rows and 8 variables.
\describe{
    \item{Date_reported}{The date in y-m-d format.}
    \item{Country_code}{Two letter country code.}
    \item{Country}{Name of the country.}
    \item{WHO_region}{World Health Organization region.}
    \item{New_cases}{new COVID-19 cases reported on that day.}
    \item{Cumulative_cases}{total number of cases reported since Jan 2020.}
    \item{New_cases}{new COVID-19 deaths reported on that day.}
    \item{Cumulative_cases}{total number of deaths reported since Jan 2020.}
}
}
\source{
\url{https://covid19.who.int/WHO-COVID-19-global-data.csv}
}
\usage{
covid_raw
}
\description{
COVID case count data taken from the world health organization. Data were
  scraped on 11-17-2023 and the updated tables from the website are subject
  to change.
}
\keyword{datasets}
