% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{crime_hawkes}
\alias{crime_hawkes}
\title{Data regarding high school completion and crime rates across U.S. states
  in 2014.}
\format{
A data.frame with 50 rows and 5 columns.
\describe{
  \item{state}{The 4-year ACGR is the number of students who graduate in 4
  years with a regular high school diploma divided by the number of students
  who form the adjusted cohort for the graduating class.” This number has
  been rounded to the nearest whole number.}
  \item{crime}{Species namRate of violent and property crimes per
  100,000 people}
  \item{crime}{Rate of violent crimes (murder, rape, robbery,
  aggravated assault) per 100,000 people}
  \item{property}{Rate of property crimes per 100,000}
}
}
\source{
\url{https://www.hawkeslearning.com/Statistics/dataset-descriptions.html#high-school-completion-and-crime-rate}
}
\usage{
crime_hawkes
}
\description{
Data regarding high school completion and crime rates across U.S. states
  in 2014.
}
\keyword{datasets}
