% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{mpg_hawkes}
\alias{mpg_hawkes}
\title{Miles Per Gallon}
\format{
A data.frame with 38,693 rows and 11 columns.
\describe{
  \item{city}{city miles per gallon}
  \item{cylinders}{number of cylinders in the vehicle engine}
  \item{displ}{engine displacement (liters)}
  \item{drive}{drive axle type (front-wheel, four-wheel, etc.)}
  \item{fuelType}{type of fuel (regular, premium, diesel, etc.)}
  \item{highway}{highway miles per gallon}
  \item{make}{manufacturer}
  \item{model}{model name}
  \item{trans}{type of transmission}
  \item{VClass}{vehicle size class}
  \item{year}{model year}
}
}
\source{
\url{https://www.seattlecentral.edu/qelp/sets/057/057.html}

\url{https://www.hawkeslearning.com/Statistics/dataset-descriptions.html#miles-per-gallon}
}
\usage{
mpg_hawkes
}
\description{
Data on vehicle fuel economy for model years 1984—2019
}
\keyword{datasets}
