% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{patient_hawkes}
\alias{patient_hawkes}
\title{Medical statistics for a sample of 100 patients from a hospital.}
\format{
A data.frame with 100 rows and 8 columns
\describe{
  \item{id}{patient id number}
  \item{gender}{patient gender (dichotomous)}
  \item{marital_status}{one of married (M), widowed (W), divorced (D) or
  single (S)}
  \item{age}{patient age}
  \item{weight}{patient weight in pounds}
  \item{cholesterol}{patient total cholesterol}
  \item{blood_pressure}{patient systolic blood pressure in mmHg}
  \item{cohort}{patients separated by decade of age}
}
}
\source{
\url{https://www.hawkeslearning.com/Statistics/beg3/dataset-descriptions.html#decreases-in-cholesterol-levels}
}
\usage{
patient_hawkes
}
\description{
Medical statistics for a sample of 100 patients from a hospital.
}
\keyword{datasets}
