% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5055_data_descriptions.R
\docType{data}
\name{pres_speech_df}
\alias{pres_speech_df}
\title{Presidential inaugural addresses}
\format{
A data.frame with 53 rows rows and 2 variables.
\describe{
    \item{year}{The year in which the inaugural address was given.}
    \item{speech}{The full text of the inaugural speech as a single
    character strring.}

}
}
\source{
\url{https://www.presidency.ucsb.edu/documents/app-categories/spoken-addresses-and-remarks/presidential/inaugural-addresses}
}
\usage{
pres_speech_df
}
\description{
This is a web scraped dataset of all presidential inauguration speeches. It
  does not include second or third inaugural addresses, nor does it include
  presidential addresses when the vice president replaced a dead president
  mid term.
}
\keyword{datasets}
