% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{satisfy_hawkes}
\alias{satisfy_hawkes}
\title{Data regarding campus crime.}
\format{
A data.frame with 15,000 rows and 12 columns.
\describe{
  \item{employee_id}{The ID number associated with the individual employee}
  \item{satisfaction_level}{How satisfied the employee is in their position
  (scale of 0 to 1)}
  \item{last_evaluation_score}{How management rated employee performance
  during the last evaluation (scale of 0 to 1)}
  \item{number_of_projects}{The number of projects an employee is currently
  working on}
  \item{average_monthly_hours}{he average number of hours the employee works
  in a month}
  \item{years_spent_at_company}{The number of years the employee has worked
  at the company}
  \item{work_accident}{A binary variable that indicates whether the employee
  experienced an accident at work}
  \item{left_company}{A binary variable that indicates whether an employee
  left the company}
  \item{promotion_in_last_5_years}{A binary variable that indicates whether
  an employee received a promotion in the last 5 years}
  \item{department}{The department that the employee works in}
  \item{salary}{The level of the employee’ s salary(low, medium, high)}
  \item{salary_range}{The dollar range for the salary levels}
}
}
\source{
\url{https://www.hawkeslearning.com/Statistics/dataset-descriptions.html#employee-satisfaction}

\url{https://www.kaggle.com/ludobenistant/hr-analytics}
}
\usage{
satisfy_hawkes
}
\description{
Data regarding campus crime.
}
\keyword{datasets}
