% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5055_data_descriptions.R
\docType{data}
\name{swe}
\alias{swe}
\title{Utah SNOTEL station SWE measurements.}
\format{
A data.frame with 32,972 rows rows and 5 variables.
\describe{
    \item{date}{The date of the measurement in y-m-d format.}
    \item{id}{The SNOTEL measurement ID.}
    \item{name}{The SNOTEL station name.}
    \item{swe}{The measured snow water equiavlent, in 10ths of millimeters.}
    \item{qc}{An NWCC quality control flag where "S" stands for suspect
    measurements.}
}
}
\source{
\url{https://wcc.sc.egov.usda.gov/reportGenerator/}
}
\usage{
swe
}
\description{
This is data obtained from the National Water and Climate Center report
  generator. It contains snow water equivalent (SWE) measurements from two
  snowpack telemetery (SNOTEL) stations in Utah. One at Tony Grove, Lake
  up Logan Canyon, and another at Ben Lomond Peak, east of Ogden.
}
\keyword{datasets}
