% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{toluca}
\alias{toluca}
\title{Comparison of lot size vs work hours at the Toluca Company}
\format{
A data.frame with 25 rows and 2 columns
\describe{
  \item{size}{lot size (no units originally provided, but
  assumed to be thousands of square feet)}
  \item{hours}{average hours it takes to make single part}
}
}
\source{
\url{http://www.cnachtsheim-text.csom.umn.edu/Kutner/Chapter\%20\%201\%20Data\%20Sets/CH01PR20.txt}
}
\usage{
toluca
}
\description{
Comparison of lot size vs work hours at the Toluca Company
}
\keyword{datasets}
